refactor: use Glide for news images
