+++ app/src/main/java/org/ole/planet/myplanet/ui/news/AdapterNews.kt
@@ -45,6 +45,14 @@ import org.ole.planet.myplanet.utilities.TimeUtils.formatDate
 import org.ole.planet.myplanet.utilities.Utilities
 import org.ole.planet.myplanet.utilities.makeExpandable
 import org.ole.planet.myplanet.ui.news.NewsActions
+import org.ole.planet.myplanet.utilities.JsonUtils
+import android.app.Dialog
+import android.graphics.Color
+import android.widget.ImageView
+import androidx.core.graphics.drawable.toDrawable
+import com.github.chrisbanes.photoview.PhotoView
+import com.bumptech.glide.Glide
+import java.io.File
 
 class AdapterNews(var context: Context, private val list: MutableList<RealmNews?>, private var currentUser: RealmUserModel?, private val parentNews: RealmNews?, private val teamName: String = "") : RecyclerView.Adapter<RecyclerView.ViewHolder?>() {
     private lateinit var rowNewsBinding: RowNewsBinding
@@ -56,7 +64,6 @@ class AdapterNews(var context: Context, private val list: MutableList<RealmNews?
     private var sharedPreferences: SharedPrefManager? = null
     private var recyclerView: RecyclerView? = null
     var user: RealmUserModel? = null
-    private var imageLoader: NewsImageLoader? = null
     private var labelManager: NewsLabelManager? = null
     private val profileDbHandler = UserProfileDbHandler(context)
     lateinit var settings: SharedPreferences
@@ -89,7 +96,6 @@ class AdapterNews(var context: Context, private val list: MutableList<RealmNews?
     fun setmRealm(mRealm: Realm?) {
         if (mRealm != null) {
             this.mRealm = mRealm
-            imageLoader = NewsImageLoader(context, this.mRealm)
             labelManager = NewsLabelManager(context, this.mRealm, currentUser)
         }
     }
@@ -100,7 +106,6 @@ class AdapterNews(var context: Context, private val list: MutableList<RealmNews?
         user = UserProfileDbHandler(context).userModel
         settings = context.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE)
         if (::mRealm.isInitialized) {
-            if (imageLoader == null) imageLoader = NewsImageLoader(context, mRealm)
             if (labelManager == null) labelManager = NewsLabelManager(context, mRealm, currentUser)
         }
         return ViewHolderNews(rowNewsBinding)
@@ -126,7 +131,7 @@ class AdapterNews(var context: Context, private val list: MutableList<RealmNews?
 
                 configureEditDeleteButtons(viewHolder, news)
 
-                imageLoader?.loadImage(viewHolder.rowNewsBinding, news)
+                loadImage(viewHolder.rowNewsBinding, news)
                 showReplyButton(viewHolder, news, position)
                 labelManager?.setupAddLabelMenu(viewHolder.rowNewsBinding, news)
                 news.let { labelManager?.showChips(viewHolder.rowNewsBinding, it) }
@@ -431,6 +436,73 @@ class AdapterNews(var context: Context, private val list: MutableList<RealmNews?
     override fun onDetachedFromRecyclerView(recyclerView: RecyclerView) {
         super.onDetachedFromRecyclerView(recyclerView)
         this.recyclerView = null
+    }
+
+    private fun loadImage(binding: RowNewsBinding, news: RealmNews?) {
+        val imageUrls = news?.imageUrls
+        if (!imageUrls.isNullOrEmpty()) {
+            try {
+                val imgObject = Gson().fromJson(imageUrls[0], JsonObject::class.java)
+                val file = File(JsonUtils.getString("imageUrl", imgObject))
+                Glide.with(context)
+                    .load(file)
+                    .placeholder(R.drawable.ic_loading)
+                    .error(R.drawable.ic_loading)
+                    .into(binding.imgNews)
+                binding.imgNews.visibility = View.VISIBLE
+                binding.imgNews.setOnClickListener {
+                    showZoomableImage(it.context, file.toString())
+                }
+                return
+            } catch (_: Exception) {
+            }
+        }
+
+        news?.imagesArray?.let { imagesArray ->
+            if (imagesArray.size() > 0) {
+                val ob = imagesArray[0]?.asJsonObject
+                val resourceId = JsonUtils.getString("resourceId", ob)
+                val library = mRealm.where(RealmMyLibrary::class.java)
+                    .equalTo("_id", resourceId)
+                    .findFirst()
+                val basePath = context.getExternalFilesDir(null)
+                val imageFile = if (library != null && basePath != null) {
+                    File(basePath, "ole/${library.id}/${library.resourceLocalAddress}")
+                } else null
+                if (imageFile?.exists() == true) {
+                    Glide.with(context)
+                        .load(imageFile)
+                        .placeholder(R.drawable.ic_loading)
+                        .error(R.drawable.ic_loading)
+                        .into(binding.imgNews)
+                    binding.imgNews.visibility = View.VISIBLE
+                    binding.imgNews.setOnClickListener {
+                        showZoomableImage(it.context, imageFile.toString())
+                    }
+                    return
+                }
+            }
+        }
+        binding.imgNews.visibility = View.GONE
+    }
+
+    private fun showZoomableImage(context: Context, imageUrl: String) {
+        val dialog = Dialog(context, android.R.style.Theme_Black_NoTitleBar_Fullscreen)
+        val view = LayoutInflater.from(context).inflate(R.layout.dialog_zoomable_image, null)
+        val photoView = view.findViewById<PhotoView>(R.id.photoView)
+        val closeButton = view.findViewById<ImageView>(R.id.closeButton)
+
+        dialog.setContentView(view)
+        dialog.window?.setBackgroundDrawable(Color.BLACK.toDrawable())
+
+        Glide.with(context)
+            .load(imageUrl)
+            .error(R.drawable.ic_loading)
+            .into(photoView)
+
+        closeButton.setOnClickListener { dialog.dismiss() }
+
+        dialog.show()
     }
 
     internal inner class ViewHolderNews(val rowNewsBinding: RowNewsBinding) : RecyclerView.ViewHolder(rowNewsBinding.root) {
